backend_bd/requirements.txt
python-dotenv
pymongo
qdrant-client
